Complete standard:connect implementation
Current code does not a way of handling a connect request to the wallet.
This commit solves this by adding state to store `on` events and a connected wallet account.

It also supports emitting the changed event to.

This commit also solves the bug where `get accounts()` was just a static array and now
a dapp can query connected walletAccount event after `on` event is emitted.

Closes #7